+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Development
 
 ### Added
+
+- heartbeat to Rebuffering state (AN-1281)
 - Data downloaded metrics to ExoPlayer
 
 ## v1.12.1